pyproject.toml
[tool.poetry]
name = "ac215-mega"
version = "0.1.0"
description = ""
authors = ["rohans0509 <59282751+rohans0509@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ffmpeg-python = "^0.2.0"
glob2 = "^0.7"
google-cloud-aiplatform = "^1.71.1"
google-cloud-speech = "^2.28.0"
google-cloud-storage = "^2.18.2"
googletrans = "^3.0.0"
requests = "^2.32.3"
vertexai = "^1.71.1"
google-cloud-texttospeech = "^2.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
